Pkg restructure: util_code.py -> code/
