+++ aiida_jutools/computer/util.py
@@ -1,3 +1,17 @@
+# -*- coding: utf-8 -*-
+###############################################################################
+# Copyright (c), Forschungszentrum Jülich GmbH, IAS-1/PGI-1, Germany.         #
+#                All rights reserved.                                         #
+# This file is part of the aiida-jutools package.                             #
+# (AiiDA JuDFT tools)                                                         #
+#                                                                             #
+# The code is hosted on GitHub at https://github.com/judftteam/aiida-jutools. #
+# For further information on the license, see the LICENSE.txt file.           #
+# For further information please visit http://judft.de/.                      #
+#                                                                             #
+###############################################################################
+"""Tools for working with aiida Computer nodes: utils."""
+
 import typing as _typing
 
 import aiida as _aiida
